requirements.txt
fastapi
uvicorn[standard]
pymupdf
pillow
pytesseract
pdfplumber
camelot-py[cv]
opencv-python-headless
numpy
pandas
scikit-learn
sentence-transformers
spacy
textblob
networkx
nltk
sumy
transformers
torch
aiofiles
python-multipart
